backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
python-multipart
llama-index
llama-index-vector-stores-azurecosmosmongo
llama-index-embeddings-openai
llama-index-core
pymongo
PyPDF2
aiofiles
fastapi-cors
